Complete:Booking process

+++ src/main/java/com/project/BookCarOnline/DTO/Response/PaymentResponse.java
@@ -15,6 +15,6 @@ public class PaymentResponse {
     String paymentUrl;
     String orderId;
     String transactionId;
-    Long amount;
+    Double amount;
     String paymentMethod; // "VNPAY" or "MOMO"
 }
